+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support of Oauth Token Federation auth for Databricks
 - Fixed new bindings syntax support for bindings that require legacy resolution and an expression in the template literal contains the token separator string (i.e. `", "`)
    - e.g. `${JSON.stringify({ name: "test", enabled: true })}`
+- Added support for "oauth on behalf of token exchange" in Salesforce integration
 
 ## v1.26.0
 - Added support for the `metadata` endpoint in the GraphQL integration